### 2021-01-04 #2934 @whedon
[PRE REVIEW]: minimo: a linked data and metadata storage system for small labs
**Submitting author:** @borchardt (<a href="http://orcid.org/0000-0001-9033-6365">Jackson Borchardt</a>)
**Repository:** <a href="https://github.com/focolab/minimo.git" target ="_blank">https://github.com/focolab/minimo.git</a>
**Version:** v1.0.0
