### 2021-01-05 #2934 @Borchardt
Thanks, majensen!

With regard to the scope check, we wanted to add a couple of notes that might provide helpful context:

Of the 3066 lines of of code reported by whedon, the 955 JSON lines (Node package information), 115 markdown lines (readme content), and 79 SVG lines (our logo) might not be considered "real" code. This still leaves 1,917 lines of core application code -- JavaScript, CSS, Dockerfile, etc.

We also wanted to note that the repo we have shared here represents a public port of work originally completed in a longer-lived private repo. That work encompassed more than a few months, and it included significant code contributions by @RaymondDunn, who is listed as an author on the paper (but whose presence is not visible in the public repo).

Thanks again; looking forward to hearing back!
