### 2021-01-09 2f3b35d
browse-data supports prefix query param; metadata rows link to browse-data
