### 2021-01-22 #2979 @whedon
[REVIEW]: minimo: a linked data and metadata storage system for small labs
**Submitting author:** @borchardt (<a href="http://orcid.org/0000-0001-9033-6365">Jackson Borchardt</a>)
**Repository:** <a href="https://github.com/focolab/minimo.git" target ="_blank">https://github.com/focolab/minimo.git</a>
**Version:** 44b306957691505bc4f95cdb0721c7e1b03a9e7d
