### 2021-03-03 #2979 @bensonml
I recommend the paper for publication

1. Software looks good - and it's great to see it being currently used by the lab.
2. installation and options worked well
3. documentation is pretty good
4. it would be nice to see some tests - even if they are smoke tests
5. look forward to seeing future releases, and mentioned updates
