### 2021-03-04 64ef47c
add test for building and connecting to web app
