+++ .github/workflows/main.yml
@@ -8,5 +8,7 @@ jobs:
       - uses: actions/checkout@v2
       - name: Build the app
         run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
+        ports:
+          - 443:443
       - name: Test connection to web app
         run: docker run --network host appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
