### 2021-03-04 beded98
try curling 127.0.0.1 instead of localhost
