+++ .github/workflows/main.yml
@@ -7,6 +7,6 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - name: Build the app
-        run: HOST_NAME=localhost docker-compose up -d -p 127.0.0.1:443:443 --build --force-recreate
+        run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Test connection to web app
-        run: docker run --network host appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
+        run: docker run --network host appropriate/curl -s --retry 10 --retry-connrefused --insecure https://127.0.0.1
