+++ .github/workflows/main.yml
@@ -9,4 +9,4 @@ jobs:
       - name: Build the app
         run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Test connection to web app
-        run: docker run --network host appropriate/curl -s --retry 10 --retry-connrefused --insecure https://127.0.0.1
+        run: docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
