### 2021-03-04 a6c0d00
try cleaning up run command now that working
