+++ .github/workflows/main.yml
@@ -7,6 +7,6 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - name: Build the app
-        run: docker-compose up -d --build --force-recreate
+        run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Test connection to web app
         run: docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
