+++ .github/workflows/main.yml
@@ -9,4 +9,10 @@ jobs:
       - name: Build the app
         run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Test connection to web app
+        id: test-connection
         run: docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
+      - uses: nick-invision@assert-action@v1
+        with:
+          expected: Welcome to the minimo FAQ!
+          actual: ${{ steps.test-connection.outputs.value }}
+          comparison: contains
