+++ .github/workflows/main.yml
@@ -10,6 +10,6 @@ jobs:
         run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Connect to web app
         id: connect
-        run: docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
+        run: echo "::set-output name=mix-ecto_create::$(docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost)"
       - name: Check landing page content
         run: echo "${{ steps.connect.outputs.value }}"
