+++ .github/workflows/main.yml
@@ -7,9 +7,15 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - name: Build the app
+        uses: ./
         run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Connect to web app
+        uses: ./
         id: connect
-        run: echo "::set-output name=mix-ecto_create::$(docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost)"
+        run: docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
       - name: Check landing page content
-        run: echo "${{ steps.connect.outputs.value }}"
+        uses: nick-invision@assert-action@v1
+        with:
+          expected: FAQ
+          actual: FAQ
+          comparison: contains
