+++ .github/workflows/main.yml
@@ -7,15 +7,10 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - name: Build the app
-        uses: ./
         run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Connect to web app
-        uses: ./
         id: connect
         run: docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
       - name: Check landing page content
         uses: nick-invision@assert-action@v1
-        with:
-          expected: FAQ
-          actual: FAQ
-          comparison: contains
+        run: echo ""
