+++ .github/workflows/main.yml
@@ -12,5 +12,4 @@ jobs:
         id: connect
         run: docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
       - name: Check landing page content
-        uses: nick-invision@assert-action@v1
         run: echo ""
