+++ .github/workflows/main.yml
@@ -10,6 +10,6 @@ jobs:
         run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Connect to web app
         id: connect
-        run: docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost
+        run: echo "::set-output name=landing_page_content::$(docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost)"
       - name: Check landing page content
         run: echo ""
