+++ .github/workflows/main.yml
@@ -14,5 +14,3 @@ jobs:
       - name: Check landing page content
         if: ${{ contains("${{ steps.curl-landing-page.outputs.landing-page-content }}", FAQ) }}
         run: echo "Found welcome header"
-        if: ${{ !contains("${{ steps.curl-landing-page.outputs.landing-page-content }}", FAQ) }}
-        run: echo "Landing page does not contain expected content" && exit 1
