+++ .github/workflows/main.yml
@@ -13,4 +13,4 @@ jobs:
         run: echo "::set-output name=landing-page-content::"$(docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost)""
       - name: Check landing page content
         if: "!contains(steps.curl-landing-page.outputs.landing-page-content, 'unexpected string')"
-        run: echo "Expected landing page content not found" && exit 1
+        run: echo "Expected landing page content not found. Instead, landing page content was:" && echo "steps.curl-landing-page.outputs.landing-page-content" && exit 1
