+++ .github/workflows/main.yml
@@ -12,5 +12,5 @@ jobs:
         id: curl-landing-page
         run: echo '::set-output name=landing-page-content::"$(docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost)"'
       - name: Check landing page content
-        if: "!contains(steps.curl-landing-page.outputs.landing-page-content, 'Welcome to the minimo FAQ!')"
+        if: ${{ (grep "Welcome to the minimo FAQ!" <<< "${{ steps.curl-landing-page.outputs.landing-page-content }}" | wc -l) != 1 }}
         run: echo -e "Expected landing page content not found. Instead, landing page content was:\n ${{ steps.curl-landing-page.outputs.landing-page-content }}" && exit 1
