+++ .github/workflows/main.yml
@@ -10,7 +10,7 @@ jobs:
         run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Get landing page content
         id: curl-landing-page
-        run: echo '::set-output name=landing-page-content::"$(docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost)"'
+        run: echo '::set-output name=landing-page-content::$(docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost)'
       - name: Find welcome header
         id: find-welcome-header
         run: echo '::set-output name=welcome-header-count::"$(grep 'Welcome to the minimo FAQ!' <<< "${{ steps.curl-landing-page.outputs.landing-page-content }}" | wc -l)"'
