+++ .github/workflows/main.yml
@@ -1,0 +1,14 @@
+# from https://github.com/peter-evans/docker-compose-actions-workflow
+name: Docker Compose Actions Workflow
+on: push
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Build the app
+        run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
+      - name: Get landing page content
+        id: curl-landing-page
+        run: echo '::set-output name=landing-page-content::$(docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost)'
+        
