### 2021-03-05 cad961a
multiline string handling
