+++ .github/workflows/main.yml
@@ -10,5 +10,7 @@ jobs:
         run: HOST_NAME=localhost docker-compose up -d --build --force-recreate
       - name: Get landing page content
         id: curl-landing-page
-        run: echo '::set-output name=landing-page-content::$(docker run --network container:minimo_reverse-proxy_1 appropriate/curl -s --retry 10 --retry-connrefused --insecure https://localhost)'
-        
+        # multiline strings require special handling in actions. see https://trstringer.com/github-actions-multiline-strings/ and https://github.com/actions/toolkit/blob/main/docs/commands.md#set-an-environment-variable
+        run: echo 'LANDING_PAGE_HTML<<EOF' >> $GITHUB_ENV && curl -s --retry 10 --retry-connrefused --insecure https://localhost >> $GITHUB_ENV && echo 'EOF' >> $GITHUB_ENV
+      - name: Print html
+        run: echo "HTML is: ${{ env.LANDING_PAGE_HTML }}"
