### 2021-03-05 44715b8
smoke test (#31)
add action to build app, connect to reverse proxy, and validate proper landing page load
