### 2021-03-05 #2979 @Borchardt
Thanks for the feedback, bensonml! I agree with you on the need for automated testing. As a quick first step, I've added an Actions workflow which builds the app, connects to its reverse proxy container, and validates landing page content.

Adding deeper functional tests is on my backlog, and I'll update here as I make headway.
