### 2021-03-08 #2979 @junjun-zhang
`minimo` certainly has the potential to address the need of a system that helps researchers to organize and track their data together with associated metadata.

Installation and getting the app up and running was straightforward, thanks to the nice documentation.

My comments are more on useability from a user’s perspective:

1. The term `form` in the `MANAGE FORMS` is a bit confusing. It seems more like a field, not a web form that consists of multiple fields.
2. If the above understanding is correct, is there any plan to add support for defining metadata `form` with multiple `fields`? That would be very helpful as I expect many labs may have different types of experimental results, different experiments can then use different metadata forms.
3. It’s not quite clear what `Number of experiments` means, if the data were generated from multiple experiments, wouldn’t it be more advisable to upload them individually for fine-grained linking between data and metadata?
4. Related to the above point, it would be nice to provide some examples of real world data and metadata that can be tried out by the users. This should give a user better ideas how she might use `minimo` to manage her own data/metadata.
5. Number of experiments set to `0` seems not allowed, why not remove `0` in the dropdown list?
6. It seems like overwriting an existing folder/file is permitted, should there be a warning before proceeding? It would be good to get confirmation from the user to prevent potential mistakes.
7. It would be good to get confirmation from the user before deleting a metadata form which has been used by many metadata entries since this operation is irreversable and possibley affects large number of entries.
8. Since there is no pagination for listing metadata entries, practically speaking, it would be difficult for the UI to scale up to over hundreds of items. Would it be possible to support listing metadata entries in a table with pagination?
9. Directly exposing MinIO UI via iframe is a quick way to allow users to access data objects managed by MinIO. However, this also comes with some drawbacks. For example, a user may bypass the normal data manipulation route to delete data files without the metadata portion being updated resulting in discrepancies. Also, a user may accidentally overwrite existing data objects without leaving any trace.
10. It would be ideal that the metadata listing table mentioned in point 8 also provides users with access to data files, eg, displaying folder structure of an uploaded entry, offering download links etc, maybe in a popup modal. This way there is no need to expose MinIO UI directly and avoid possible caveats.
11. When a metadata entry is updated using `link existing data with a new metadata entry`, the previous metadata entry still shows up as any others, could this be confusing? If it’s desirable to keep revision history, it would be nice to collapse the two metadata entries with the latest shown by default.
12. I once tried to set up a user with a password, but later forgot the password. It seems there is no password reset functionality, I ended up locking myself out, had to create a new deployment.
