### 2021-03-11 focolab/minimo#32 @Borchardt
usability: metadata forms rather than metadata fields
Now that we're considering supporting larger labs/organizations, showing all fields to all users may get out of hand quickly; user-defined collections of fields more manageable.
