### 2021-03-11 focolab/minimo#33 @Borchardt
usability: "number of experiments"
It's not clear whether this text in the upload flow makes sense for the general use case. We need to come to some consensus and update accordingly.
