### 2021-03-11 focolab/minimo#34 @Borchardt
docs: examples of real/more interesting metadata forms
Possibilities / intended use cases should be more obvious to users before installation.
