### 2021-03-11 focolab/minimo#35 @Borchardt
feature: disable or discourage overwrites and deletions
They're currently enabled, but we'd prefer that they weren't. Ideally, disable them through MinIO configuration; at a minimum, warn users before carrying them out.
