### 2021-03-11 focolab/minimo#36 @Borchardt
usability: warn before delete frequently used metadata forms
Something along the lines of "You are deleting a metadata form which has been used `m` times by `n` users. Please click 'okay' to proceed."
