### 2021-03-11 focolab/minimo#37 @Borchardt
feature: metadata browser allows direct raw data downloads
Rather than just linking to the data browser, the metadata browser should expose batch download ability directly.
