### 2021-03-11 #2979 @Borchardt
Thanks for your feedback, junjun-zhang. Comments/responses below:

1. Good point; agree. Will be fixed by solution to point 2.
2. We think that this is a good suggestion. I've created [an issue in the tracker](https://github.com/focolab/minimo/issues/32) for this work.
3. We agree that the text here probably needs to change, but we're still discussing what alternative might make the most sense. I've created [an issue in the tracker](https://github.com/focolab/minimo/issues/33) to track this work.
4. We think this sounds like a reasonable update for the docs. I've created [an issue in the tracker](https://github.com/focolab/minimo/issues/34) for this work.
5. Good point, and related to [existing issue #13](https://github.com/focolab/minimo/issues/13) in our issue tracker. I've updated that issue to include removing 0 from the dropdown.
6. Overwrites are currently permitted, we'd prefer that it weren't. I've created [an issue in the tracker](https://github.com/focolab/minimo/issues/35) for disabling them.
7. I don't know that this is a high priority, but I do think that it's a sensible idea. I've created [an issue for the associated work](https://github.com/focolab/minimo/issues/36)
8. Agreed. We have [an existing issue in our tracker](https://github.com/focolab/minimo/issues/10) for this.
9. For the time being (until point 10 is resolved), exposing the MinIO browser is necessary in order to provide the ability for downloads of raw data folders. Some of the reviewer's concerns here may be mitigated once work is completed for point 6, but we are open to moving away from exposing the MinIO browser whenever we no longer have a functional need to do so.
10. Agreed. I've created [an issue to track associated work](https://github.com/focolab/minimo/issues/37).
11. Our point of view here is that maintaining an "inseparable link" between data and metadata means that old (even incorrect) metadata should not be deleted nor hidden. We believe that timestamping of metadata entries provides sufficient versioning, at least for the time being.
12. We appreciate this feedback. I think that it should be possible for anyone with root access to the host machine to wipe the users database without nuking the entire installation; I've created [an issue for adding instructions for doing so](https://github.com/focolab/minimo/issues/38) in our tracker.
