### 2021-03-17 9f22aee
front end and back end blocking of illegal object names (#41)
