### 2021-03-26 0068980
equality queries working
