+++ server.js
@@ -548,7 +548,10 @@ app.get('/browse-datametadata', checkAuth, (req, res) => {
           metadata_fields.push(forms[i]['element name']);
         }
 
-        let mongo_query = (req.query && req.query.search_field && req.query.search_term ? {[req.query.search_field]: req.query.search_term} : {});
+        let mongo_query = {};
+        if (req.query && req.query.search_field) {
+          mongo_query[req.query.search_field] = ( req.query.search_term ? { $regex: req.query.search_term } : { $exists: true } );
+        }
 
         // get all form data
         docs = app.locals.DB.getDocuments(mongoConfig.metadataCollection, mongo_query)
