### 2021-03-26 ee7b308
Issue 10 (#45)
* add search functionality to metadata browser

* sort metadata by descending upload date
