### 2021-04-03 #2979 @majensen
@Borchardt the paper text looks fine. Can you add examples of the sizes of data objects that you use with minimo now (after "supports very large files")?

I'd like to see some more work towards the usability issues identified by @junjun-zhang. The following tickets are open from your response to Junjun above:

https://github.com/focolab/minimo/issues/32
https://github.com/focolab/minimo/issues/33
https://github.com/focolab/minimo/issues/34
https://github.com/focolab/minimo/issues/35
https://github.com/focolab/minimo/issues/36
https://github.com/focolab/minimo/issues/37
https://github.com/focolab/minimo/issues/38

From these, I think 32, 35, 36, and 38 would be highest priority. What do you think, could you implement these or a subset?
thanks
