### 2021-04-03 #2979 @Borchardt
@majensen Thanks for your work!

Work for https://github.com/focolab/minimo/issues/38 was actually merged a couple of weeks ago, but I apparently forgot to link and close the issue. I've done so now.

Of the remaining three you highlighted, I think focolab/minimo#35 is probably the highest priority. I'm beginning work for that issue now, and I will aim to have a PR for it in the next couple of days. From there, I'll transition to 32 and then 36. Our lab would prefer that we not block publication for those last two issues, but I will aim to have them completed within the next week regardless.

How does all of that sound?
