### 2021-04-05 d3f2b4b
add mc container and make it create readonly user
user account gets correct permissions

update environment variable names to reflect addition of separate minio user account

update readme for new minimo user password flag
