### 2021-04-06 14a8b9c
Issue 35 (#46)
* force unique prefixes for uploads

* data browser uses read only credentials
