+++ server.js
@@ -10,7 +10,10 @@ const mongoose = require('mongoose');
 const passport = require('passport');
 const LocalStrategy = require('passport-local').Strategy;
 const Account = require('./models/account');
-const { v4: uuidv4 } = require('uuid');
+const {
+  v1: uuidv1,
+  v4: uuidv4
+} = require('uuid');
 
 // for minio
 const Minio = require('minio');
@@ -18,8 +21,8 @@ const Minio = require('minio');
 const minioClient = new Minio.Client({
   endPoint: process.env.HOST_NAME,
   useSSL: false,
-  accessKey: process.env.MINIO_ACCESS_KEY,
-  secretKey: process.env.MINIO_SECRET_KEY,
+  accessKey: process.env.MINIO_ADMIN_USERNAME,
+  secretKey: process.env.MINIO_ADMIN_PASSWORD,
   region: process.env.MINIO_REGION
 });
 
@@ -28,8 +31,8 @@ const DB = require('./javascripts/mongo-db');
 const mongoConfig = require('./elements/config-mongo');
 const formElementStrings = require('./javascripts/form-element-strings');
 const formatMetadataEntriesForDisplay = require('./javascripts/format-metadata-entries-display');
-// per https://docs.aws.amazon.com/AmazonS3/latest/userguide/object-keys.html
-const valid_object_name = /^[a-z0-9\/!\-_\.\*'\(\)]{1,1024}$/i;
+// per https://docs.aws.amazon.com/AmazonS3/latest/userguide/object-keys.html; max length of 987 rather than 1024 because of autogenerated 37 character prefix (36 char uuid + slash)
+const valid_object_name = /^[a-z0-9\/!\-_\.\*'\(\)]{1,987}$/i;
 
 // javascripts
 const validateMetadataForm = require('./javascripts/validate-metadata-form');
@@ -308,23 +311,31 @@ app.get('/submit-data-boxuploader', checkAuth, (req, res) => {
   }
 });
 
-// taken from https://docs.min.io/docs/upload-files-from-browser-using-pre-signed-urls.html
-app.get('/presigned-url', checkAuth, (req, res) => {
-  if (valid_object_name.test(req.query.name)) {
+// based on https://docs.min.io/docs/upload-files-from-browser-using-pre-signed-urls.html
+app.get('/presigned-urls', checkAuth, (req, res) => {
+  let presignedUrls = [];
+  let filenames = req.query.filenames.split(',');
+  let prefix = uuidv1(); // use v1 to guarantee unique prefix
 
-    minioClient.presignedPutObject('data', req.query.name, (err, url) => {
-      if (err) {
-        errmsg = `Error while trying to put object. \n`;
-        handleError(req, res, err, errmsg);
-      } else {
-        res.end(url);
-      }
-    });
+  for (var i = 0; i < filenames.length; i++) {
+    
+    if (valid_object_name.test(filenames[i])) {
+      let object_key = prefix + '/' + filenames[i];
+      minioClient.presignedPutObject('data', object_key, (err, url) => {
+        if (err) {
+          errmsg = `Error while trying to put object. \n`;
+          handleError(req, res, err, errmsg);
+        } else {
+          presignedUrls.push(url);
+        }
+      });
 
-  } else {
-    errmsg = `Attempted to upload object with invalid name. Only alphanumeric characters (a-z, A-Z, and 0-9) and the special characters /, !, -, _, ., *, ', (, and ) are permitted in object names. Object names must consist of at least 1 character and no more than 1024 characters.\n`;
-    handleError(req, res, null, errmsg);
+    } else {
[…]
