+++ README.md
@@ -152,7 +152,7 @@ minimo is a [Node](https://nodejs.org) web server. This server talks to [MinIO](
 
 This architecture offers several advantages over other options that we considered:
 
-1. **It supports very large files.** MinIO, which we use for raw data storage, supports file sizes up to 5TB. Since our lab deals primarily in volumetric microscope videos, which can grow quite large, this ability was important to us.
+1. **It supports very large files.** MinIO, which we use for raw data storage, supports file sizes up to 5TB, and we have successfully uploaded files as large as 25GiB through our web interface. Since our lab deals primarily in volumetric microscope videos, which can grow quite large, this ability was important to us.
 2. **It provides both easy human access and robust programmatic access to its data stores.** We hope that our web interface, which abstracts away the technical details of data and metadata storage, will allow users to access and upload data with little overhead. At the same time, users interested in accessing data programmatically can do so with MinIO's and MongoDB's SDKs.
 3. **Each of its components can be deployed both on local hardware and in the cloud.** The app's Node server can be run either on a local machine or on a cloud server. It supports both local and cloud hosted MongoDB endpoints for metadata storage. MinIO is compatible with AWS S3, so it is possible use S3 in place of local hardware for data storage and/or backup. This gives users a large degree of control over the costs and operational loads they will incur and over the performance they will achieve.
 4. **It is lightweight enough to be easily extensible.** We intend to continue to add functionality to the app, and we want to allow and encourage our users to do the same!
