### 2021-04-06 #2979 @Borchardt
@majensen quick update: https://github.com/focolab/minimo/issues/35 has been completed, and I've updated the paper per your request.
