### 2021-04-07 98a38e8
remove unnecessary log writes
