### 2021-04-07 45b566a
add etcd container and connect to minio (#47)
* add etcd container and connect to minio

* fix presigned-urls race
