+++ server.js
@@ -293,10 +293,8 @@ app.get('/submit-data-boxuploader', checkAuth, (req, res) => {
 
         // grab returned docs
         docs = resolve;
-        // console.log('docs0: ' + JSON.stringify(docs[0]))
 
         // add docs to locals
-        console.log(JSON.stringify(docs));
         res.locals.docs = JSON.stringify(docs);
         res.locals.username = getUsername(req);
         res.locals.valid_object_name = valid_object_name;
@@ -312,7 +310,7 @@ app.get('/submit-data-boxuploader', checkAuth, (req, res) => {
 });
 
 // based on https://docs.min.io/docs/upload-files-from-browser-using-pre-signed-urls.html
-app.get('/presigned-urls', checkAuth, (req, res) => {
+app.get('/presigned-urls', checkAuth, async (req, res) => {
   let presignedUrls = [];
   let filenames = req.query.filenames.split(',');
   let prefix = uuidv1(); // use v1 to guarantee unique prefix
@@ -321,14 +319,13 @@ app.get('/presigned-urls', checkAuth, (req, res) => {
     
     if (valid_object_name.test(filenames[i])) {
       let object_key = prefix + '/' + filenames[i];
-      minioClient.presignedPutObject('data', object_key, (err, url) => {
-        if (err) {
-          errmsg = `Error while trying to put object. \n`;
-          handleError(req, res, err, errmsg);
-        } else {
-          presignedUrls.push(url);
-        }
-      });
+      try {
+        const url = await minioClient.presignedPutObject('data', object_key);
+        presignedUrls.push(url);
+      } catch (err) {
+        errmsg = `Error while trying to put object. \n`;
+        handleError(req, res, err, errmsg);
+      }
 
     } else {
       errmsg = `Attempted to upload object with invalid name. Only alphanumeric characters (a-z, A-Z, and 0-9) and the special characters /, !, -, _, ., *, ', (, and ) are permitted in object names. Object names must consist of at least 1 character and no more than 1024 characters.\n`;
@@ -378,7 +375,6 @@ app.get('/manage-forms', checkAuth, (req, res) => {
 
         // grab returned docs
         docs = resolve;
-        // console.log('docs0: ' + JSON.stringify(docs[0]))
 
         // add docs to locals
         res.locals.docs = JSON.stringify(docs);
@@ -528,10 +524,8 @@ app.get('/link-metadata-to-data', checkAuth, (req, res) => {
 
         // grab returned docs
         docs = resolve;
-        // console.log('docs0: ' + JSON.stringify(docs[0]))
 
         // add docs to locals
-        console.log(JSON.stringify(docs));
         res.locals.docs = JSON.stringify(docs);
         res.locals.username = getUsername(req);
 
@@ -571,7 +565,6 @@ app.get('/browse-datametadata', checkAuth, (req, res) => {
 
             // grab returned docs
             docs = resolve;
-            console.log(`docs0: ${JSON.stringify(docs[0])}`);
 
             // get formatted docs
             formattedDocs = formatMetadataEntriesForDisplay(docs, metadata_fields);
