### 2021-04-07 1a16a93
delete more unnecessary console logs
