+++ javascripts/form-element-strings.js
@@ -2,17 +2,20 @@
 function formElementObjToString(formJSONObj) {
   // assuming one object
   id = formJSONObj._id;
-  name = formJSONObj['element name'];
+  form_name = formJSONObj['form name'];
   elements = formJSONObj['elements'];;
-
-  // format string
-  // double escape to render properly in html
-  formattedString = `${name}\\n`
   
+  formattedString = `${form_name}\\n`
+
   for (i = 0; i < elements.length; i++) {
+    formattedString += '\\n'
+    name = elements[i]['element name'];
     prompt = elements[i].prompt;
     type = elements[i]['element type'];
-    formattedString += `${prompt}\\n${type}\\n`;
+
+    // format string
+    // double escape to render properly in html
+    formattedString += `${name}\\n${prompt}\\n${type}\\n`;
 
     // append options if it's a select
     if (type == 'dropdown') {
@@ -36,34 +39,26 @@ function formElementStringToObj(formJSONString) {
 
   // replace possible line breaks with regular one
   fmtString = formJSONString.replace(/(\r\n|\r)/gm, '\n');
-  // split on return
-  splitString = fmtString.split('\n');
+  // split on double newline
+  splitString = fmtString.split('\n\n');
 
-  // check to make sure type is in right place
-  typeNdx = 2;
-  if (!(types.indexOf(splitString[typeNdx]) > -1)) {
-    errmsg = 'Form text box not formatted correctly!';
-    throw errmsg;
-  }
-
-  name = splitString[0];
+  form_name = splitString[0];
   elements = splitString.slice(1);
   obj = {
-    'element name': name,
+    'form name': form_name,
     'elements': [],
   };
 
-  let i=0;
-  while (i < elements.length) {
-    let element_type = elements[i];
+  for (i = 0; i < elements.length; i++) {
+    lines = elements[i].split('\n');
+
+    let element_type = lines[0];
     if (!(types.indexOf(element_type) > -1)) {
       errmsg = 'Form text box not formatted correctly!';
       throw errmsg;
     }
-    i += 1;
 
-    prompt = elements[i];
-    i += 1;
+    prompt = lines[1];
 
     let element = {
       'element type': element_type,
@@ -72,19 +67,17 @@ function formElementStringToObj(formJSONString) {
 
     if ('dropdown' === element_type) {
       let options = [];
-      while (i < elements.length && '' !== elements[i]) {
+      for (j = 2; j < lines.length; j++) {
         // remove leading/trailing whitespace
-        trimmed = elements[i].trim();
+        trimmed = lines[j].trim();
         if (trimmed.length > 0) {
           options.push(trimmed);
         }
-        i += 1;
       }
       element.options = options;
     }
 
     obj.elements.push(element);
-    i += 1;
   }
 
   return obj;
