### 2021-04-10 #2979 @majensen
Ok @Borchardt, last thing we need from you is an open archive of the code submitted to Zenodo or similar. There is a little Zenodo tutorial at [this link](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729). JOSS prefers that the archive title match the paper title, and the archive authors match the paper authors.
Once you have this, can you report the version number and the archive DOI back here? Thanks!
