### 2021-04-10 #2979 @Borchardt
Awesome! I'll get on this as soon as I've merged the outstanding PR for https://github.com/focolab/minimo/issues/36.

Just double checking: there's no need for me to merge code changes from mainline into our (now outdated) joss branch, correct?
