### 2021-04-10 44b3069
Issue 36 (#49)
* track form submission counts

* get confirmation before deleting previously used forms
